+++ tests/Rank2.hs
@@ -2,10 +2,10 @@ module Rank2(main) where
 import Prelude
 
 f :: (forall a . a -> a) -> (Int, Bool)
-f i = (i 1, i True)
+f i = (i (1::Int), i True)
 
 g :: (forall a . a -> Int -> a) -> (Int, Bool)
-g c = (c 1 1, c True 1)
+g c = (c (1::Int) (1::Int), c True (1::Int))
 
 data Id = Id (forall a . a -> a)
 
@@ -17,4 +17,4 @@ main = do
   putStrLn $ show $ f id
   putStrLn $ show $ g const
   case iD of
-    Id i -> putStrLn $ show (i 1, i True)
+    Id i -> putStrLn $ show (i (1::Int), i True)
